+++ packages/plugins/io-center/src/store.svelte.ts
@@ -7,7 +7,7 @@ import {
 } from "./ied/object-tree.type.d";
 import type { Nullable } from "./types";
 import type { Connection, LogicalConditioner } from "./ui/components/canvas/types.canvas";
-import type { LpElement } from "./ui/components/lp-list/types.lp-list";
+import type { LpElement } from "./ui/components/right-bar/lp-list/types.lp-list";
 
 
 export class Store {
@@ -34,9 +34,30 @@ export class Store {
 	dataObjects: DataObject[] = $state([])
 	objectTree: ObjectTree = $state(NullObjectTree)
 
+	//
+	// #region Logical Conditioners
+	//
 	logicalConditioners: LogicalConditioner[] = $state([])
+	_selectedLogicalConditioners: LogicalConditioner[] = $state([])
+	selectedLogicalConditioners: LogicalConditioner[] = $derived([...this._selectedLogicalConditioners])
 	public findLC(type: string, instance: string) {
 		return this.logicalConditioners.find(lc => lc.type === type && lc.instance === instance)
+	}
+	public selectLC(lc: LogicalConditioner) {
+		this._selectedLogicalConditioners.push(lc)
+	}
+	public deselectLC(lc: LogicalConditioner) {
+		this._selectedLogicalConditioners = this._selectedLogicalConditioners.filter((item) => item.id !== lc.id);
+	}
+	public isLcSelected(lc: LogicalConditioner) {
+		return this._selectedLogicalConditioners.some((selectedLC) => selectedLC.id === lc.id)
+	}
+	public toggleLcSelection(lc: LogicalConditioner) {
+		if (this.isLcSelected(lc)) {
+			this.deselectLC(lc);
+		} else {
+			this.selectLC(lc);
+		}
 	}
 
 
@@ -44,8 +65,8 @@ export class Store {
 	public resetConnections(_: unknown) {
 		this.connections = []
 	}
-	public connectionExistsFor(lp: LpElement) {
-		return this.connections.some(connection => connection.to.name.includes(`${lp.type}-${lp.instance}`))
+	public connectionExistsFor(element: LpElement | LogicalConditioner) {
+		return this.connections.some(connection => connection.to.name.includes(`${element.type}-${element.instance}`))
 	}
 
 	// 
@@ -66,7 +87,7 @@ export class Store {
 	public isLPSelected(lp: LpElement) {
 		return this._selectedLogicalPhysicals.some((selectedLP) => selectedLP.id === lp.id)
 	}
-	public toggleElementSelection(element: LpElement) {
+	public toggleLpElementSelection(element: LpElement) {
 		if (this.isLPSelected(element)) {
 			this.deselectLP(element);
 		} else {